@@ -1,5 +1,5 @@
-module github.com/prozz/aws-embedded-metrics-golang
+module github.com/gaeste/aws-embedded-metrics-golang
 
-go 1.14
+go 1.20
 
 require github.com/kinbiko/jsonassert v1.1.1
